requirements.txt
Flask
gevent-websocket
flask-sockets
git+https://github.com/openai/whisper.git
gevent
ffmpeg-python
noisereduces
